+++ lib/tools/batch/documents.py
@@ -91,19 +91,6 @@ class RevitDocumentOpener(object):
             self.ui_application.DialogBoxShowing -= self._dismiss_coordination_model_load_error
 
 
-def save_sync_and_relinquish(document, comment):
-    """Persist a changed batch model and relinquish all owned workshared data."""
-    if document.IsWorkshared:
-        transact_options = DB.TransactWithCentralOptions()
-        relinquish_options = DB.RelinquishOptions(True)
-        sync_options = DB.SynchronizeWithCentralOptions()
-        sync_options.SetRelinquishOptions(relinquish_options)
-        sync_options.Comment = comment
-        document.SynchronizeWithCentral(transact_options, sync_options)
-        DB.WorksharingUtils.RelinquishOwnership(
-            document, relinquish_options, transact_options)
-    else:
-        document.Save()
 
 
 class OpenedBatchDocument(object):
