+++ pyArchitect.tab/Batch.panel/Navis.pushbutton/script.py
@@ -1,33 +1,8 @@
 # -*- coding: utf-8 -*-
 
-import os
-import sys
-
-# ==========================================================
-# LIB
-# ==========================================================
-
-script_dir = os.path.dirname(__file__)
-
-lib_dir = os.path.join(
-    script_dir,
-    "lib"
-)
-
-if lib_dir not in sys.path:
-
-    sys.path.append(lib_dir)
-
-# ==========================================================
-# IMPORTS
-# ==========================================================
-
-from ui import show_form
+from tools.navis.ui import show_form
 from tools.navis.settings import configure
-
-from process_models import (
-    process_models
-)
+from tools.navis.process_models import process_models
 
 # ==========================================================
 # START
